+++ server/src/nest/oauth/oauth.module.ts
@@ -8,9 +8,13 @@ import { AddonsModule } from '../addons/addons.module';
 
 /**
  * OAuth 2.1 server (MCP). Public token/userinfo/revoke endpoints + the SPA's
- * authenticated consent/client/session management. The SDK-mounted
- * /oauth/authorize, /oauth/register and /oauth/consent stay on Express, so the
- * strangler lists /oauth/token, /oauth/userinfo, /oauth/revoke explicitly.
+ * authenticated consent/client/session management. /oauth/authorize,
+ * /oauth/register and /oauth/consent are still served by the MCP SDK's own
+ * router, mounted on the Express platform before app.init() — that path
+ * reaches the domain through oauth.bridge.ts, which is why the pending
+ * authorization codes live in a module, not on the provider.
+ *
+ * No `exports`: nothing inside the container injects this domain.
  */
 @Module({
   imports: [RateLimitModule, AuditModule, AddonsModule],
